+++ gestao-direcao/app/gerenciamentoAlunos/page.module.css
@@ -1,364 +1,33 @@
-/* 
-.text{
-  justify-content: center;
-  align-items: center;
-  text-align: center;
+.container {
+  text-align: center;  /* Centraliza o texto dentro do container */
+  margin-top: 50px;  /* Adiciona um espaçamento superior de 50px */
 }
 
-.filtro{
-
-  display: flex;
-
-  justify-content: space-around;
-
-  width: 100%;
-
-  align-items: center;
-
-  flex-direction: row;
-
-  margin-top: 3%;
-
-  margin-bottom: 3%;
-}
-
-.button {
-
-  display: flex;
-
-  margin-right: 5%;
-
-  background-color: #878787;
+.links {
+  display: flex;  /* Usa Flexbox para alinhar os itens dentro da div */
+  justify-content: center;  /* Centraliza os itens horizontalmente */
+  gap: 40px;  /* Define um espaçamento de 40px entre os itens */
+  margin-top: 40px;  /* Adiciona um espaçamento superior de 40px */
   
-  border-radius: 5%;
-
-  padding: 10px;
-
-  border-radius: 10px;
 
 
 }
 
-.button:hover {
-
-  background-color: #a3a1a1;
-
-  cursor: pointer;
-
-}
-
-.ano {
-
-  display: flex;
-
-  background-color: #878787;
-
-  padding: 10px;
-
-  border-radius: 10px;
-
-  justify-content: space-around;
-
-  width: 15%;
-}
-
-.input {
-
-  border-radius: 5px;
-
-  border-color: solid black 2px;
-
-  text-align: center;
-
-  background-color: #acacac;
-
+.linkItem p {
+  text-align: center;  /* Centraliza o conteúdo de texto dentro de cada link */
+  cursor: pointer;  /* Altera o cursor para "pointer" (mãozinha), indicando que o item é clicável */
+  color: rgb(0, 0, 0);  /* Define a cor do texto como preto */
+  text-decoration: none; 
+  
 }
 
 
-
-.botao {
-
-  margin-bottom: 6.7%;
-
-  display: flex;
-
-  justify-content: space-around;
-
+.linkItem a {
+  text-decoration: none; /* Remove o sublinhado do link */
 }
 
-.table {
-
-  width: 100%;
-  border-collapse: collapse;
-  margin: 20px 0;
-  font-size: 16px;
-  text-align: left;
-}
-
-.th, .td {
-  padding: 10px;
-  border: 1px solid #000000;
-}
-
-.th {
-  background-color: #f4f4f4;
-  border-bottom: 2px solid #000000;
-}
-
-.td {
-  border-bottom: 1px solid #000000;
-}
-
-.tr:nth-child(even) {
-  background-color: #f9f9f9;
-}
-
-.tr:hover {
-  background-color: #f1f1f1;
-}
-
-@media only screen and (max-width: 800px) {
-  .filtro {
-    flex-direction: row;
-    
-    flex-wrap: wrap;
-    
-    width: 80%;
-
-    display: flex;
-
-    justify-content: space-around;
-
-    margin-left: 5%;
-  }
-
-  .button {
-    
-    justify-content: space-around;
-    
-    margin: 6.7%;
-  }
-  .ano {
-    width: 100px;
-  }
-  .input {
-    width: 100%;
-  }
-  .botao {
-    flex-direction: row;
-    width: 100%;
-    margin-bottom: 9%;
-  }
-  .table {
-    font-size: 14px;
-  }
-
-}
- */
-
- .text{
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-}
-
-
-.filtro{
-
-
-  display: flex;
-
-
-  justify-content: space-around;
-
-
-  width: 100%;
-
-
-  align-items: center;
-
-
-  flex-direction: row;
-
-
-  margin-top: 3%;
-
-
-  margin-bottom: 3%;
-}
-
-
-.button {
-
-
-  display: flex;
-
-
-  margin-right: 5%;
-
-
-  background-color: #878787;
+.icon {
+  margin-bottom: 10px;  /* Adiciona um espaçamento inferior de 10px abaixo do ícone */
+  color:red;  /* Define a cor do ícone como azul (#0070f3) */
  
-  border-radius: 5%;
-
-
-  padding: 10px;
-
-
-  border-radius: 10px;
-
-
-
-
-}
-
-
-.button:hover {
-
-
-  background-color: #a3a1a1;
-  cursor: pointer;
-
-
-}
-
-
-.ano {
-
-
-  display: flex;
-
-
-  background-color: #878787;
-
-
-  padding: 10px;
-
-
-  border-radius: 10px;
-
-
-  justify-content: space-around;
-
-
-  width: 15%;
-}
-
-
-.input {
-
-
-  border-radius: 5px;
-
-
-  border-color: solid black 2px;
-
-
-  text-align: center;
-
-
-  background-color: #acacac;
-
-
-}
-
-
-
-
-
-
-.botao {
-
-
-  margin-bottom: 6.7%;
-
-
-  display: flex;
-
-
-  justify-content: space-around;
-
-
-}
-.table {
-
-
-  width: 100%;
-  border-collapse: collapse;
-  margin: 0 auto;
-  font-size: 16px;
-  text-align: center;
-  margin-top: 3%;
-}
-
-
-.table th, .table td {
-  padding: 10px;
-  border: 1px solid #000000;
-}
-
-
-.table th {
-  background-color: #f4f4f4;
-  border-bottom: 2px solid #000000;
-}
-
-
-.table td {
-  border-bottom: 1px solid #000000;
-}
-
-
-.footer {
-  position: relative;
-  bottom: 0;
-  width: 100%;
-}
-
-
-
-
-
-
-@media only screen and (max-width: 800px) {
-  .filtro {
-    flex-direction: row;
-   
-    flex-wrap: wrap;
-   
-    width: 80%;
-
-
-    display: flex;
-    justify-content: space-around;
-
-
-    margin-left: 5%;
-  }
-
-
-  .button {
-   
-    justify-content: space-around;
-   
-    margin: 6.7%;
-  }
-  .ano {
-    width: 100px;
-  }
-  .input {
-    width: 100%;
-  }
-  .botao {
-    flex-direction: row;
-    width: 100%;
-    margin-bottom: 9%;
-  }
-  .table {
-    font-size: 7px;
-    width: auto;
-    height: auto;
-  }
-
-
 }
