+++ gestao-direcao/app/gerenciamentoAlunos/users/create/createUser.module.css
@@ -1,0 +1,25 @@
+/* app/users/create/createUser.module.css */
+
+.container {
+    margin: 20px;
+  }
+  
+  .form {
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+  }
+  
+  .input {
+    padding: 8px;
+    border: 1px solid #878787;
+  }
+  
+  .button {
+    padding: 10px;
+    background-color: #878787;
+    color: black;
+    border: none;
+    cursor: pointer;
+    border-radius: 5px;
+  }
