+++ gestao-direcao/app/gerenciamentoAlunos/users/page.module.css
@@ -1,0 +1,218 @@
+
+.text{
+    justify-content: center;
+    align-items: center;
+    text-align: center;
+  }
+  
+  
+  .filtro{
+  
+  
+    display: flex;
+  
+  
+    justify-content: space-around;
+  
+  
+    width: 100%;
+  
+  
+    align-items: center;
+  
+  
+    flex-direction: row;
+  
+  
+    margin-top: 3%;
+  
+  
+    margin-bottom: 3%;
+  }
+  
+  
+  .button {
+  
+  
+    display: flex;
+  
+  
+    margin-right: 5%;
+  
+  
+    background-color: #878787;
+   
+    border-radius: 5%;
+  
+  
+    padding: 10px;
+  
+  
+    border-radius: 10px;
+  
+  
+  
+  
+  }
+  
+  
+  .button:hover {
+  
+  
+    background-color: #a3a1a1;
+    cursor: pointer;
+  
+  
+  }
+  
+  
+  .ano {
+  
+  
+    display: flex;
+  
+  
+    background-color: #878787;
+  
+  
+    padding: 10px;
+  
+  
+    border-radius: 10px;
+  
+  
+    justify-content: space-around;
+  
+  
+    width: 15%;
+  }
+  
+  
+  .input {
+  
+  
+    border-radius: 5px;
+  
+  
+    border-color: solid black 2px;
+  
+  
+    text-align: center;
+  
+  
+    background-color: #acacac;
+  
+  
+  }
+  
+  .botao {
+  
+  
+    margin-bottom: 6.7%;
+  
+  
+    display: flex;
+  
+  
+    justify-content: space-around;
+  
+  
+  }
+  .table {
+  
+  
+    width: 100%;
+    border-collapse: collapse;
+    margin: 0 auto;
+    font-size: 16px;
+    text-align: center;
+    margin-top: 3%;
+  }
+  
+  
+  .table th, .table td {
+    padding: 10px;
+    border: 1px solid #000000;
+  }
+  
+  
+  .table th {
+    background-color: #f4f4f4;
+    border-bottom: 2px solid #000000;
+  }
+  
+  
+  .table td {
+    border-bottom: 1px solid #000000;
+  }
+  
+  
+  .footer {
+    position: relative;
+    bottom: 0;
+    width: 100%;
+  }
+  
+  .novo {
+    display: flex;
+    flex-direction: row;
+    background-color: #878787;
+    color: #000000;
+    border-radius: 8px;
+  
+  }
+  
+  .novo:hover {
+    background-color: #a3a1a1;
+    cursor: pointer;
+  
+  }
+  
+  
+  
+  
+  
+  
+  @media only screen and (max-width: 800px) {
+    .filtro {
+      flex-direction: row;
+     
+      flex-wrap: wrap;
+     
+      width: 80%;
+  
+  
+      display: flex;
+      justify-content: space-around;
+  
+  
+      margin-left: 5%;
+    }
+  
+  
+    .button {
+     
+      justify-content: space-around;
+     
+      margin: 6.7%;
+    }
+    .ano {
+      width: 100px;
+    }
+    .input {
+      width: 100%;
+    }
+    .botao {
+      flex-direction: row;
+      width: 100%;
+      margin-bottom: 9%;
+    }
+    .table {
+      font-size: 7px;
+      width: auto;
+      height: auto;
+    }
+  
+  
+  }
+
