+++ gestao-direcao/package.json
@@ -9,9 +9,10 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "next": "14.2.13",
     "react": "^18",
     "react-dom": "^18",
-    "next": "14.2.13"
+    "react-icons": "^5.3.0"
   },
   "devDependencies": {
     "eslint": "^8",
